refactor(storage)!: P6.3-5 move layers + key_provider; migrate tests
Physical move of layer/{encryption,cache,audit,scope}.rs +
layer/key_provider.rs from nebula-credential to nebula-storage
(ADR-0029 Section 2, amended by ADR-0032 — impls move, trait stays).

Section 12.5 AAD binding preserved bit-for-bit per ADR-0028 invariant 1.
Pre/post grep of AAD-related lines in encryption.rs shows zero
semantic change — only line numbers shift because the `use crate::{...}`
block was restructured into `use nebula_credential::{...}` + a
`use super::super::key_provider::KeyProvider;` sibling import.

Migrated integration tests:
- tests/env_provider.rs -> tests/credential_env_provider.rs (storage)
- tests/units/encryption_tests.rs -> tests/credential_encryption_invariants.rs (storage)

Credential retains CredentialStore trait + DTOs in store.rs (ADR-0032).
Intra-doc links to layer::EncryptionLayer in store.rs / crypto.rs
replaced with plain backticks (type no longer reachable from credential).
Removed `pub mod layer;` + flat re-exports of layer + KeyProvider symbols
from `nebula-credential`; the root-doc 'Key types' section now points
readers to `nebula_storage::credential` per ADR-0032.

Embedded tests in storage layer files (and `CountingKeyProvider` /
`FailingKeyProvider` helpers in `key_provider.rs`) are gated on the
`test-util` feature so clippy/check without features skip them —
`test_helpers` in credential is itself feature-gated, and storage's
`test-util` forwards `nebula-credential/test-util`.

Storage cargo-shear ignored list trimmed to (aws-config, aws-sdk-s3,
redis, tracing, uuid) — the crypto / moka / parking_lot / scopeguard
deps are now materially used by the moved files.

Added tempfile to storage [dev-dependencies] for FileKeyProvider tests.

Ref: docs/adr/0032-credential-store-canonical-home.md
Ref: docs/adr/0028-cross-crate-credential-invariants.md Section 1 (invariant 1)
