+++ crates/storage/src/credential/layer/mod.rs
@@ -1,8 +1,11 @@
-//! Composable storage layers for [`CredentialStore`](super::CredentialStore).
+//! Composable storage layers for [`CredentialStore`](nebula_credential::CredentialStore).
 
 pub mod audit;
 pub mod cache;
 pub mod encryption;
 pub mod scope;
 
-// TODO(P6.4, P6.5): re-export once files are populated.
+pub use audit::{AuditEvent, AuditLayer, AuditOperation, AuditResult, AuditSink};
+pub use cache::{CacheConfig, CacheLayer, CacheStats};
+pub use encryption::EncryptionLayer;
+pub use scope::{ScopeLayer, ScopeResolver};
