--- crates/storage/src/credential/store.rs
@@ -1,3 +1,0 @@
-//! Placeholder — real implementation ported in P6.2.
-//!
-//! See docs/adr/0029-storage-owns-credential-persistence.md §2.
