+++ crates/storage/tests/credential_encryption_invariants.rs
